Updated Simulation
Made the simulation actually do the method of least squares estimation.
I had to create a loop that augmented the matrices correctly, by stacking
the corret timed values on top of each other. Then you do the right
pseudoinverse to solve the equation. You must use the \ operator instead of
the inverse function. The inverse funtion does not accept doubles as
parameters.